+++ client/src/components/Chat/Messages/Content/ToolCallGroup.tsx
@@ -24,6 +24,7 @@ import { isError, StackedToolIcons } from './ToolOutput';
 import { isBashProgrammaticToolCall } from './routing';
 import { ASK_USER_QUESTION } from '~/utils/approval';
 import { useMCPIconMap, useMCPServerNames } from '~/hooks/MCP';
+import SearchVerticals from './verticals';
 import { ROW_GLYPH_SLOT } from './rows';
 import store from '~/store';
 
@@ -565,7 +566,10 @@ export default function ToolCallGroup({
         )}
       </div>
       {groupAttachments && groupAttachments.length > 0 && (
-        <AttachmentGroup attachments={groupAttachments} />
+        <>
+          <SearchVerticals attachments={groupAttachments} />
+          <AttachmentGroup attachments={groupAttachments} />
+        </>
       )}
     </div>
   );
